+++ benefits/src/sdk/models/operations/getbenefitsplanenrollment.ts
@@ -23,7 +23,7 @@ export class GetBenefitsPlanEnrollmentResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * List of plan enrollments for a member