+++ benefits/src/sdk/models/operations/patchbenefitsserviceid.ts
@@ -40,5 +40,5 @@ export class PatchBenefitsServiceIdResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }