+++ benefits/src/sdk/sdk.ts
@@ -3,11 +3,11 @@
  */
 
 import * as utils from "../internal/utils";
-import * as errors from "../sdk/models/errors";
-import * as operations from "../sdk/models/operations";
 import * as shared from "../sdk/models/shared";
+import { BenefitsEnrollment } from "./benefitsenrollment";
+import { BenefitsService } from "./benefitsservice";
 import axios from "axios";
-import { AxiosInstance, AxiosRequestConfig, AxiosResponse, RawAxiosRequestHeaders } from "axios";
+import { AxiosInstance } from "axios";
 
 /**
  * Contains the list of servers available to the SDK
@@ -27,6 +27,8 @@ export const ServerList = [
  * The available configuration options for the SDK
  */
 export type SDKProps = {
+    bearerAuth?: string;
+
     /**
      * Allows overriding the default axios client used by the SDK
      */
@@ -49,13 +51,14 @@ export type SDKProps = {
 
 export class SDKConfiguration {
     defaultClient: AxiosInstance;
+    security?: shared.Security | (() => Promise<shared.Security>);
     serverURL: string;
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "1.0.0";
-    sdkVersion = "2.0.0";
-    genVersion = "2.188.1";
-    userAgent = "speakeasy-sdk/typescript 2.0.0 2.188.1 1.0.0 @wingspan/benefits";
+    sdkVersion = "2.0.1";
+    genVersion = "2.194.1";
+    userAgent = "speakeasy-sdk/typescript 2.0.1 2.194.1 1.0.0 @wingspan/benefits";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);
@@ -66,6 +69,15 @@ export class SDKConfiguration {
  * Wingspan Benefits API: Benefits
  */
 export class Benefits {
+    /**
+     * Operations related to enrollments
+     */
+    public benefitsEnrollment: BenefitsEnrollment;
+    /**
+     * Operations related to service management
+     */
+    public benefitsService: BenefitsService;
+
     private sdkConfiguration: SDKConfiguration;
 
     constructor(props?: SDKProps) {
@@ -79,351 +91,13 @@ export class Benefits {
         const defaultClient = props?.defaultClient ?? axios.create();
         this.sdkConfiguration = new SDKConfiguration({
             defaultClient: defaultClient,
+            security: new shared.Security({ bearerAuth: props?.bearerAuth }),
+
             serverURL: serverURL,
             retryConfig: props?.retryConfig,
         });
-    }
 
-    /**
-     * Retrieve Enrollment Details for a Specific Member
-     *
-     * @remarks
-     * Fetches the enrollment status and details for a member identified by the provided unique identifier.
-     */
-    async getBenefitsEnrollmentId(
-        req: operations.GetBenefitsEnrollmentIdRequest,
-        config?: AxiosRequestConfig
-    ): Promise<operations.GetBenefitsEnrollmentIdResponse> {
-        if (!(req instanceof utils.SpeakeasyBase)) {
-            req = new operations.GetBenefitsEnrollmentIdRequest(req);
-        }
-
-        const baseURL: string = utils.templateUrl(
-            this.sdkConfiguration.serverURL,
-            this.sdkConfiguration.serverDefaults
-        );
-        const operationUrl: string = utils.generateURL(baseURL, "/benefits/enrollment/{id}", req);
-        const client: AxiosInstance = this.sdkConfiguration.defaultClient;
-        const headers: RawAxiosRequestHeaders = { ...config?.headers };
-        headers["Accept"] = "application/json";
-
-        headers["user-agent"] = this.sdkConfiguration.userAgent;
-
-        const httpRes: AxiosResponse = await client.request({
-            validateStatus: () => true,
-            url: operationUrl,
-            method: "get",
-            headers: headers,
-            responseType: "arraybuffer",
-            ...config,
-        });
-
-        const responseContentType: string = httpRes?.headers?.["content-type"] ?? "";
-
-        if (httpRes?.status == null) {
-            throw new Error(`status code not found in response: ${httpRes}`);
-        }
-
-        const res: operations.GetBenefitsEnrollmentIdResponse =
-            new operations.GetBenefitsEnrollmentIdResponse({
-                statusCode: httpRes.status,
-                contentType: responseContentType,
-                rawResponse: httpRes,
-            });
-        const decodedRes = new TextDecoder().decode(httpRes?.data);
-        switch (true) {
-            case httpRes?.status == 200:
-                if (utils.matchContentType(responseContentType, `application/json`)) {
-                    res.enrollment = utils.objectToClass(JSON.parse(decodedRes), shared.Enrollment);
-                } else {
-                    throw new errors.SDKError(
-                        "unknown content-type received: " + responseContentType,
-                        httpRes.status,
-                        decodedRes,
-                        httpRes
-                    );
-                }
-                break;
-        }
-
-        return res;
-    }
-
-    /**
-     * List all plan enrollments
-     */
-    async getBenefitsPlanEnrollment(
-        config?: AxiosRequestConfig
-    ): Promise<operations.GetBenefitsPlanEnrollmentResponse> {
-        const baseURL: string = utils.templateUrl(
-            this.sdkConfiguration.serverURL,
-            this.sdkConfiguration.serverDefaults
-        );
-        const operationUrl: string = baseURL.replace(/\/$/, "") + "/benefits/plan-enrollment";
-        const client: AxiosInstance = this.sdkConfiguration.defaultClient;
-        const headers: RawAxiosRequestHeaders = { ...config?.headers };
-        headers["Accept"] = "application/json";
-
-        headers["user-agent"] = this.sdkConfiguration.userAgent;
-
-        const httpRes: AxiosResponse = await client.request({
-            validateStatus: () => true,
-            url: operationUrl,
-            method: "get",
-            headers: headers,
-            responseType: "arraybuffer",
-            ...config,
-        });
-
-        const responseContentType: string = httpRes?.headers?.["content-type"] ?? "";
-
-        if (httpRes?.status == null) {
-            throw new Error(`status code not found in response: ${httpRes}`);
-        }
-
-        const res: operations.GetBenefitsPlanEnrollmentResponse =
-            new operations.GetBenefitsPlanEnrollmentResponse({
-                statusCode: httpRes.status,
-                contentType: responseContentType,
-                rawResponse: httpRes,
-            });
-        const decodedRes = new TextDecoder().decode(httpRes?.data);
-        switch (true) {
-            case httpRes?.status == 200:
-                if (utils.matchContentType(responseContentType, `application/json`)) {
-                    res.classes = [];
-                    const resFieldDepth: number = utils.getResFieldDepth(res);
-                    res.classes = utils.objectToClass(
-                        JSON.parse(decodedRes),
-                        shared.PlanEnrollment,
-                        resFieldDepth
-                    );
-                } else {
-                    throw new errors.SDKError(
-                        "unknown content-type received: " + responseContentType,
-                        httpRes.status,
-                        decodedRes,
-                        httpRes
-                    );
-                }
-                break;
-        }
-
-        return res;
-    }
-
-    /**
-     * Get a particular plan enrollment by ID
-     */
-    async getBenefitsPlanEnrollmentId(
-        req: operations.GetBenefitsPlanEnrollmentIdRequest,
-        config?: AxiosRequestConfig
-    ): Promise<operations.GetBenefitsPlanEnrollmentIdResponse> {
-        if (!(req instanceof utils.SpeakeasyBase)) {
-            req = new operations.GetBenefitsPlanEnrollmentIdRequest(req);
-        }
-
-        const baseURL: string = utils.templateUrl(
-            this.sdkConfiguration.serverURL,
-            this.sdkConfiguration.serverDefaults
-        );
-        const operationUrl: string = utils.generateURL(
-            baseURL,
-            "/benefits/plan-enrollment/{id}",
-            req
-        );
-        const client: AxiosInstance = this.sdkConfiguration.defaultClient;
-        const headers: RawAxiosRequestHeaders = { ...config?.headers };
-        headers["Accept"] = "application/json";
-
-        headers["user-agent"] = this.sdkConfiguration.userAgent;
-
-        const httpRes: AxiosResponse = await client.request({
-            validateStatus: () => true,
-            url: operationUrl,
-            method: "get",
-            headers: headers,
-            responseType: "arraybuffer",
-            ...config,
-        });
-
-        const responseContentType: string = httpRes?.headers?.["content-type"] ?? "";
-
-        if (httpRes?.status == null) {
-            throw new Error(`status code not found in response: ${httpRes}`);
-        }
-
-        const res: operations.GetBenefitsPlanEnrollmentIdResponse =
-            new operations.GetBenefitsPlanEnrollmentIdResponse({
-                statusCode: httpRes.status,
-                contentType: responseContentType,
-                rawResponse: httpRes,
-            });
-        const decodedRes = new TextDecoder().decode(httpRes?.data);
-        switch (true) {
-            case httpRes?.status == 200:
-                if (utils.matchContentType(responseContentType, `application/json`)) {
-                    res.planEnrollment = utils.objectToClass(
-                        JSON.parse(decodedRes),
-                        shared.PlanEnrollment
-                    );
-                } else {
-                    throw new errors.SDKError(
-                        "unknown content-type received: " + responseContentType,
-                        httpRes.status,
-                        decodedRes,
-                        httpRes
-                    );
-                }
-                break;
-        }
-
-        return res;
-    }
-
-    /**
-     * Retrieve Current Benefits Service Status
-     *
-     * @remarks
-     * Fetches the current status indicating whether the benefits service is enabled or disabled.
-     */
-    async getBenefitsService(
-        config?: AxiosRequestConfig
-    ): Promise<operations.GetBenefitsServiceResponse> {
-        const baseURL: string = utils.templateUrl(
-            this.sdkConfiguration.serverURL,
-            this.sdkConfiguration.serverDefaults
-        );
-        const operationUrl: string = baseURL.replace(/\/$/, "") + "/benefits/service";
-        const client: AxiosInstance = this.sdkConfiguration.defaultClient;
-        const headers: RawAxiosRequestHeaders = { ...config?.headers };
-        headers["Accept"] = "application/json";
-
-        headers["user-agent"] = this.sdkConfiguration.userAgent;
-
-        const httpRes: AxiosResponse = await client.request({
-            validateStatus: () => true,
-            url: operationUrl,
-            method: "get",
-            headers: headers,
-            responseType: "arraybuffer",
-            ...config,
-        });
-
-        const responseContentType: string = httpRes?.headers?.["content-type"] ?? "";
-
-        if (httpRes?.status == null) {
-            throw new Error(`status code not found in response: ${httpRes}`);
-        }
-
-        const res: operations.GetBenefitsServiceResponse =
-            new operations.GetBenefitsServiceResponse({
-                statusCode: httpRes.status,
-                contentType: responseContentType,
-                rawResponse: httpRes,
-            });
-        const decodedRes = new TextDecoder().decode(httpRes?.data);
-        switch (true) {
-            case httpRes?.status == 200:
-                if (utils.matchContentType(responseContentType, `application/json`)) {
-                    res.serviceEnablementResponse = utils.objectToClass(
-                        JSON.parse(decodedRes),
-                        shared.ServiceEnablementResponse
-                    );
-                } else {
-                    throw new errors.SDKError(
-                        "unknown content-type received: " + responseContentType,
-                        httpRes.status,
-                        decodedRes,
-                        httpRes
-                    );
-                }
-                break;
-        }
-
-        return res;
-    }
-
-    /**
-     * Modify Benefits Service Status
-     *
-     * @remarks
-     * Allows users to change the enablement status of a specified benefits service.
-     */
-    async patchBenefitsServiceId(
-        req: operations.PatchBenefitsServiceIdRequest,
-        config?: AxiosRequestConfig
-    ): Promise<operations.PatchBenefitsServiceIdResponse> {
-        if (!(req instanceof utils.SpeakeasyBase)) {
-            req = new operations.PatchBenefitsServiceIdRequest(req);
-        }
-
-        const baseURL: string = utils.templateUrl(
-            this.sdkConfiguration.serverURL,
-            this.sdkConfiguration.serverDefaults
-        );
-        const operationUrl: string = utils.generateURL(baseURL, "/benefits/service/{id}", req);
-
-        let [reqBodyHeaders, reqBody]: [object, any] = [{}, null];
-
-        try {
-            [reqBodyHeaders, reqBody] = utils.serializeRequestBody(
-                req,
-                "serviceEnablementUpdate",
-                "json"
-            );
-        } catch (e: unknown) {
-            if (e instanceof Error) {
-                throw new Error(`Error serializing request body, cause: ${e.message}`);
-            }
-        }
-        const client: AxiosInstance = this.sdkConfiguration.defaultClient;
-        const headers: RawAxiosRequestHeaders = { ...reqBodyHeaders, ...config?.headers };
-        headers["Accept"] = "application/json";
-
-        headers["user-agent"] = this.sdkConfiguration.userAgent;
-
-        const httpRes: AxiosResponse = await client.request({
-            validateStatus: () => true,
-            url: operationUrl,
-            method: "patch",
-            headers: headers,
-            responseType: "arraybuffer",
-            data: reqBody,
-            ...config,
-        });
-
-        const responseContentType: string = httpRes?.headers?.["content-type"] ?? "";
-
-        if (httpRes?.status == null) {
-            throw new Error(`status code not found in response: ${httpRes}`);
-        }
-
-        const res: operations.PatchBenefitsServiceIdResponse =
-            new operations.PatchBenefitsServiceIdResponse({
-                statusCode: httpRes.status,
-                contentType: responseContentType,
-                rawResponse: httpRes,
-            });
-        const decodedRes = new TextDecoder().decode(httpRes?.data);
-        switch (true) {
-            case httpRes?.status == 200:
-                if (utils.matchContentType(responseContentType, `application/json`)) {
-                    res.serviceEnablementResponse = utils.objectToClass(
-                        JSON.parse(decodedRes),
-                        shared.ServiceEnablementResponse
-                    );
-                } else {
-                    throw new errors.SDKError(
-                        "unknown content-type received: " + responseContentType,
-                        httpRes.status,
-                        decodedRes,
-                        httpRes
-                    );
-                }
-                break;
-        }
-
-        return res;
+        this.benefitsEnrollment = new BenefitsEnrollment(this.sdkConfiguration);
+        this.benefitsService = new BenefitsService(this.sdkConfiguration);
     }
 }