+++ benefits/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [typescript v2.0.0] benefits
 ### Releases
 - [NPM v2.0.0] https://www.npmjs.com/package/@wingspan/benefits/v/2.0.0 - benefits
+
+## 2023-11-20 01:26:15
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.0.1] benefits
+### Releases
+- [NPM v2.0.1] https://www.npmjs.com/package/@wingspan/benefits/v/2.0.1 - benefits